requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.2.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
PyJWT==2.6.0
python-dotenv==1.0.0
pytz==2023.3
requests==2.28.2
twilio==7.17.0
urllib3==1.26.15
Werkzeug==2.2.3
